Changed rampDuration to float

+++ Sources/CAudioKit/include/AKGlobals.hpp
@@ -9,14 +9,14 @@
 
 extern "C" float __akDefaultSampleRate;
 extern "C" int __akDefaultChannelCount;
-extern "C" double __akRampDuration;
+extern "C" float __akRampDuration;
 extern "C" bool __akEnableEchoCancellation;
 
 #else
 
 extern float __akDefaultSampleRate;
 extern int __akDefaultChannelCount;
-extern double __akRampDuration;
+extern float __akRampDuration;
 extern bool __akEnableEchoCancellation;
 
 #endif
